requirements.txt
# UI и графический интерфейс
PyQt5==5.15.10

# Базовые утилиты и переменные окружения
python-dotenv==1.0.1

# Мониторинг файлов и системы
watchdog==4.0.0
psutil==5.9.8

# Модуль камеры (Фото-ловушка)
opencv-python==4.9.0.80

# Telegram-бот и алерты
pyTelegramBotAPI==4.15.4
requests==2.31.0

# Генерация отчетов (PDF)
reportlab==4.1.0

# Для будущей сборки в .exe (чтобы на защите не через консоль запускать)
pyinstaller==6.4.0

